backend/requirements.txt
requests
python-dotenv
pyTelegramBotAPI
pytz
Django
